backend/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
websockets>=12.0
ollama>=0.4.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
python-multipart>=0.0.9
